requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.8.6
requests==2.26.0